[E2E] Consolidate question [WIP] (#24733)
* Remove test that is not relevant anymore

* Move browse data test where it belongs

* Move repro 16108 to the `sharing` reproductions

* Consolidate tests regarding the new question data picker

* Remove duplicate test

Exactly the same copy lives in `saved.cy.spec.js`

* Move tests related to aggregation to where they belong

* Migrate data picker reproductions over from `new-menu` spec

* Move test to the saved questions

* Remove the `describe` block related to "simple" question mode

Simple or Custom question modes do not exist anymore.
When opening a new question data picker, it's always in the custom/notebook mode.

The new "simple" mode now refers to an already opened table or a saved question.

* Move repro 15808 to nested questions
